feat: SignUpTemplate
